+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@ Authentication Service MKII release notes
 0.4.1
 -----
 
+* Added a `/testmode` endpoint that mimics the standard root endpoint.
 * The service is now tested against OpenJDK 8 and 11.
   * Note that the compiler compliance level must still be set at 1.8. The server fails to
     start if the compliance level is 11, likely due to out of date jars, including, possibly,